+++ community/plugins/spring-ai-alibaba-starter-plugin-larksuite/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.alibaba.cloud.ai.plugin.config.LarksuiteAutoConfiguration
+com.alibaba.cloud.ai.plugin.larksuite.LarksuiteAutoConfiguration
